backend/requirements.txt
Flask
pandas
kafka-python
pybreaker
psycopg2-binary
sqlalchemy
